+++ src/main/java/cart/config/CartWebConfiguration.java
@@ -2,6 +2,7 @@ package cart.config;
 
 import cart.auth.AuthArgumentResolver;
 import cart.auth.LoginInterceptor;
+import cart.service.MembersService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -13,9 +14,15 @@ import java.util.List;
 @Configuration
 public class CartWebConfiguration implements WebMvcConfigurer {
 
+    private final MembersService membersService;
+
+    public CartWebConfiguration(MembersService membersService) {
+        this.membersService = membersService;
+    }
+
     @Bean
     public LoginInterceptor loginInterceptor() {
-        return new LoginInterceptor();
+        return new LoginInterceptor(membersService);
     }
 
     @Override
